seperate the lvglTask
